Fix to tropopause height calculation and additional debug output to pseudo-map files
The tropopause height calculation now tries to avoid accidentally putting it in the
boundary layer, which can happen if extrapolation causes the potential temperature
to cross the tropopause value in the PBL. The algorithm now requires that the
tropopause be in the part of the profile where the potential temperature increases
with altitude.

Added the date that the CO2 concentration from MLO/SMO was taken from for each level
and the width of any averaging used.

Finally, added an option to switch between putting the CO2 profiles on the GGG2014
0:1:70 km grid or the native GEOS grid.