Quick patch to remove the extreme seasonal cycle in N2O and CH4 in extrapolation backward
Testing showed that, since CH4 and N2O needed to be extrapolated
backward in time 10-20 years, the seasonal cycle became grossly
amplified by the extrapolation. The easiest way to fix this was
just to apply a 12 month rolling mean to the back extrapolated
part of all trace gas records. This is not the ideal way (both
because we lose any seasonality and it assumes a linear decrease
over the time period in question) but will do for now.