Better netCDF attribution
hg_is_commit_clean now allows you to specify files to ignore, this
is important for writing the creation message into the netCDF file
since if the netCDF file is tracked, the commit will always be dirty
when writing to that file. Also added ability to store SHA1 hashes
of input files as netCDF attributes so we can always be sure which
file(s) produced the output file.