Modified how strat LUT dependencies are handled to include code dependencies.
This means likely abandoning the idea of tracking the LUTs and distributing
with the code. Which is fine, I think its more important that we ensure the
LUTs are recalculated if the code changes.  There's no fancy dependency
analysis, the modules it depends on need to be manually listed, which allows
us to ignore any environmental packages.