Cleaned up vertical grid code for tccon_priors
Removed an old option to keep the priors on a GEOS grid
or regrid to the 0:1:70 km grid. For now it is not needed
because gsetup.f handles the regridding as needed. In the
future, it can be added with more flexibility to read a
grid from an integral file. An integral reading function
has been added to mod_utils for that purpose.