Fix for Feb 29th bug in tropospheric prior
For the tropospheric "equivalent latitude", we rely on a lookup table of
theta in the mid-troposphere vs. latitude to capture synoptic shifts of
airmasses that make the profiles have a different shape than would be
captured by their geographic latitude. This LUT was constructed from 2018
data, so it lacks Feb 29th. To allow profiles to be created for Feb 29th,
we just say it should look like Feb 28th.