Fix to prior flags in acos_interface's parallel mode
The way that multiprocessing works, there's no way to easily
do shared arrays, so the helper functions running in parallel
needed to return the flags, and the parallel function needed
to stitch them back together.

Also fixed the error handler to properly limit the error
messages when not running verbosely.