Crucial bugfix: non-3D variables in native files were being flipped
In mod_maker with native level GEOS files, the variables need to
be flipped vertically to change them from being ordered space-to-surface
to surface-to-space. However, the logic assumed that all variables
read in from the profile files were 3D, but the PHIS variable is
not. Consequently, its latitude dimension was getting flipped resulting
in incorrect surface altitudes. This has now been corrected.